+++ exonum-light-client/src/main/java/com/exonum/client/ExplorerApiHelper.java
@@ -28,7 +28,8 @@ import com.exonum.binding.common.message.TransactionMessage;
 import com.exonum.client.response.Block;
 import com.exonum.client.response.BlockResponse;
 import com.exonum.client.response.BlocksResponse;
-import com.exonum.client.response.ServiceInfo;
+import com.exonum.client.response.ServiceInstanceInfo;
+import com.exonum.client.response.ServiceInstanceState;
 import com.exonum.client.response.ServicesResponse;
 import com.exonum.client.response.TransactionResponse;
 import com.exonum.client.response.TransactionStatus;
@@ -93,9 +94,11 @@ final class ExplorerApiHelper {
     );
   }
 
-  static List<ServiceInfo> parseServicesResponse(String json) {
+  static List<ServiceInstanceInfo> parseServicesResponse(String json) {
     ServicesResponse servicesResponse = JSON.fromJson(json, ServicesResponse.class);
-    return servicesResponse.getServices();
+    return servicesResponse.getServices().stream()
+        .map(ServiceInstanceState::getSpec)
+        .collect(toList());
   }
 
   private static ExecutionStatus getExecutionStatus(